DGD208-Spring2025-BeraSamilKilinc-FINAL
This version is the final version of this GitHub repo. Below are a few notes:

-Hunger increases by 1 every 0.5 seconds.

-Sleep decreases by 1 every 4 seconds.

-Fun decreases by 1 every 2 seconds.

-Every time Fun is used, the character gets tired, so Sleep decreases by 10.

-If the character sleeps too much, since too much time has passed, the pet’s hunger will increase, so be careful!

(When using an item, the screen refreshes regularly, so the menu input you typed may appear to be erased, but it is not actually erased. For example, if you selected option 2 from the menu and the number 2 you typed seems to disappear, usually just pressing Enter is enough.)

Note: Pets are saved in the pets.json file located in the PetGame folder under Roaming.